requirements.txt
aiofiles>=24.1.0
aiohappyeyeballs>=2.4.6
aiohttp>=3.11.12
aiosignal>=1.3.2
async-timeout>=5.0.1
asyncpg>=0.30.0
attrs>=25.1.0
colorama>=0.4.6
discord.py>=2.5.0
frozenlist>=1.5.0
idna>=3.10
loguru>=0.7.3
multidict>=6.1.0
pip>=25.0.1
python-dotenv>=1.0.1
yarl>=1.18.3
